Intern modularization of the CategoryForm
The Category Form, especifically the CreateFieldForm have been dividen
in:
- CreateTextField
- CreateListField
- CreateFieldButtons
The functions and states provided to the components are passed via
Zustand, with the useCategoryForm, anyways the components uses its own
state for manaign the controlled inputs.
